Исправление согласно требованиям проверки lab_2.py

+++ lab_2.py
@@ -1,18 +1,23 @@
-from icrawler.builtin import GoogleImageCrawler
+import argparse
 import os
 import csv
-import argparse
+
+from icrawler.builtin import GoogleImageCrawler
+from datetime import datetime
 
 
 def download_images(filename_imagesdata: str) -> None:
     """
     Скачивание картинок
     """
+    year = datetime.now().year
+    if os.path.exists(filename_imagesdata) == 0:
+        os.mkdir(filename_imagesdata)
     while len(os.listdir(filename_imagesdata)) < 50:
         Google_crawler = GoogleImageCrawler(storage={"root_dir": filename_imagesdata})
         Google_crawler.crawl(
             keyword="monkey",
-            filters=dict(date=((2025, 1, 1), (2025, 12, 31))),
+            filters=dict(date=((year, 1, 1), (year, 12, 31))),
             max_num=1000,
         )
 
@@ -21,27 +26,31 @@ def make_annotation_file(filename_annotation: str, filename_imagesdata: str) -> 
     """
     Создание и запись файла аннотации
     """
-    with open(filename_annotation, mode="w", newline="", encoding="utf-8") as file:
-        writer = csv.writer(file)
-        for f in os.listdir(filename_imagesdata):
-            path_full = os.path.abspath(f)
-            path = os.path.join(filename_imagesdata, f)
-            writer.writerow([path_full, path])
+    if os.path.exists(filename_imagesdata):
+        with open(filename_annotation, mode="w", newline="", encoding="utf-8") as file:
+            writer = csv.writer(file)
+            writer.writerow(["Absolute Path", "Relative Path"])
+            for f in os.listdir(filename_imagesdata):
+                path = os.path.join(filename_imagesdata, f)
+                path_full = os.path.abspath(path)
+                writer.writerow([path_full, path])
 
 
 class Path_Iterator:
-    def __init__(self, source):
+    def __init__(self, source: str):
+        self.items = []
+        self.counter = 0
         if os.path.isfile(source):
-            with open(source, newline='', encoding='utf-8') as file:
+            with open(source, newline="", encoding="utf-8") as file:
                 reader = csv.reader(file)
-                self.items = [row for row in reader]
-                self.counter = 0
+                next(reader)
+                for row in reader:
+                    self.items.append(row)
         else:
-            for f in source:
-                path_full = os.path.abspath(f)
-                path = os.path.join(source, f)
-                self.items.append([path_full,path])
-            self.counter = 0
+            for file in os.listdir(source):
+                path = os.path.join(source, file)
+                path_full = os.path.abspath(path)
+                self.items.append([path_full, path])
 
     def __iter__(self):
         return self
@@ -61,8 +70,7 @@ def parsing() -> tuple[str, str]:
     """
     parser = argparse.ArgumentParser()
     parser.add_argument("filenameimagesdata", type=str)
-    parser.add_argument(
-        "filenameannotation", type=str)
+    parser.add_argument("filenameannotation", type=str)
     args = parser.parse_args()
     return args.filenameimagesdata, args.filenameannotation
 
@@ -80,4 +88,5 @@ def main() -> None:
 
 if __name__ == "__main__":
     main()
+
